refactor: Section 분리
